+++ sites/public/pages/eligibility/bedrooms.tsx
@@ -9,11 +9,13 @@ import {
   t,
   Form,
   FieldGroup,
+  ProgressNav,
 } from "@bloom-housing/ui-components"
 import FormsLayout from "../../layouts/forms"
 import { useForm } from "react-hook-form"
 import React from "react"
 import { useRouter } from "next/router"
+import { ELIGIBILITY_ROUTE, ELIGIBILITY_SECTIONS } from "../../lib/constants"
 
 const EligibilityBedrooms = () => {
   const router = useRouter()
@@ -34,6 +36,13 @@ const EligibilityBedrooms = () => {
 
   return (
     <FormsLayout>
+      <FormCard header={t("eligibility.progress.header")}>
+        <ProgressNav
+          currentPageSection={2}
+          completedSections={1}
+          labels={ELIGIBILITY_SECTIONS.map((label) => t(`eligibility.progress.sections.${label}`))}
+        />
+      </FormCard>
       <FormCard>
         <Form onSubmit={handleSubmit(onSubmit)}>
           <div className="form-card__group is-borderless">
@@ -53,7 +62,7 @@ const EligibilityBedrooms = () => {
             <div className="form-card__pager-row primary">
               <Button
                 styleType={AppearanceStyleType.primary}
-                onClick={() => router.push("/eligibility/age")}
+                onClick={() => router.push(`/${ELIGIBILITY_ROUTE}/${ELIGIBILITY_SECTIONS[2]}`)}
               >
                 {t("t.next")}
               </Button>