+++ sites/public/pages/eligibility/welcome.tsx
@@ -2,11 +2,19 @@
 Eligibility Welcome
 Explanation about the purpose of the questionnaire.
 */
-import { AppearanceStyleType, Button, FormCard, t, Form } from "@bloom-housing/ui-components"
+import {
+  AppearanceStyleType,
+  Button,
+  FormCard,
+  t,
+  Form,
+  ProgressNav,
+} from "@bloom-housing/ui-components"
 import FormsLayout from "../../layouts/forms"
 import { useForm } from "react-hook-form"
 import React from "react"
 import { useRouter } from "next/router"
+import { ELIGIBILITY_ROUTE, ELIGIBILITY_SECTIONS } from "../../lib/constants"
 
 const EligibilityWelcome = () => {
   const router = useRouter()
@@ -19,6 +27,13 @@ const EligibilityWelcome = () => {
 
   return (
     <FormsLayout>
+      <FormCard header={t("eligibility.progress.header")}>
+        <ProgressNav
+          currentPageSection={1}
+          completedSections={0}
+          labels={ELIGIBILITY_SECTIONS.map((label) => t(`eligibility.progress.sections.${label}`))}
+        />
+      </FormCard>
       <FormCard>
         <div className="form-card__lead border-b">
           <h2 className="form-card__title is-borderless mt-4">{t("eligibility.welcome.header")}</h2>
@@ -31,7 +46,7 @@ const EligibilityWelcome = () => {
             <div className="form-card__pager-row primary">
               <Button
                 styleType={AppearanceStyleType.primary}
-                onClick={() => router.push("/eligibility/bedrooms")}
+                onClick={() => router.push(`/${ELIGIBILITY_ROUTE}/${ELIGIBILITY_SECTIONS[1]}`)}
               >
                 {t("t.next")}
               </Button>